gradio-demo/requirements.txt
# Python 3.12 환경
gradio==4.12.0
ultralytics==8.0.0
pillow==10.1.0
opencv-python==4.8.1.78
